WIP fix conversion type mismatches
