+++ codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/smithyrust/generator/RustLibraryShimGenerator.java
@@ -11,6 +11,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
+import software.amazon.polymorph.traits.DafnyUtf8BytesTrait;
 import software.amazon.polymorph.traits.LocalServiceTrait;
 import software.amazon.polymorph.utils.IOUtils;
 import software.amazon.polymorph.utils.MapUtils;
@@ -493,18 +494,11 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
 
   private String structureGetter(final MemberShape memberShape) {
     final Map<String, String> variables = memberVariables(memberShape);
-
-    // for some simple shapes, the Rust runtime types are not Copy
-    final Shape targetShape = model.expectShape(memberShape.getTarget());
-    final boolean needsClone =
-      targetShape.isBlobShape() || targetShape.isStringShape();
-    variables.put("fieldClone", needsClone ? ".clone()" : "");
-
     final String template =
       """
       #[allow(missing_docs)] // documentation missing in model
-      pub fn $fieldName:L(&self) -> ::std::option::Option<$fieldType:L> {
-          self.$fieldName:L$fieldClone:L
+      pub fn $fieldName:L(&self) -> &::std::option::Option<$fieldType:L> {
+          &self.$fieldName:L
       }
       """;
     return IOUtils.evalTemplate(template, variables);
@@ -753,9 +747,9 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
     );
     final Map<String, String> variables = MapUtils.merge(
       serviceVariables(),
-      operationVariables(operationShape)
+      operationVariables(operationShape),
+      structureVariables(structureShape)
     );
-    variables.put("structureName", structureId.getName(service));
     variables.put(
       "variants",
       toDafnyVariantsForStructure(structureShape).toString()
@@ -766,7 +760,7 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
         """
         #[allow(dead_code)]
         pub fn to_dafny(
-            value: crate::operation::$snakeCaseOperationName:L::$structureName:L,
+            value: crate::operation::$snakeCaseOperationName:L::$rustStructureName:L,
         ) -> ::std::rc::Rc<
             crate::r#$dafnyTypesModuleName:L::$structureName:L,
         >{
@@ -810,9 +804,9 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
     );
     final Map<String, String> variables = MapUtils.merge(
       serviceVariables(),
-      operationVariables(operationShape)
+      operationVariables(operationShape),
+      structureVariables(structureShape)
     );
-    variables.put("structureName", structureId.getName(service));
     variables.put(
       "fluentMemberSetters",
       fluentMemberSettersForStructure(structureShape).toString()
@@ -827,8 +821,8 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
             dafny_value: ::std::rc::Rc<
                 crate::r#$dafnyTypesModuleName:L::$structureName:L,
             >,
-        ) -> crate::operation::$snakeCaseOperationName:L::$structureName:L {
-            crate::operation::$snakeCaseOperationName:L::$structureName:L::builder()
+        ) -> crate::operation::$snakeCaseOperationName:L::$rustStructureName:L {
+            crate::operation::$snakeCaseOperationName:L::$rustStructureName:L::builder()
                 $fluentMemberSetters:L
                 .build()
                 .unwrap()
@@ -1001,6 +995,295 @@ public class RustLibraryShimGenerator extends AbstractRustShimGenerator {
       // TODO: union
       default -> throw new UnsupportedOperationException(
         "Unsupported shape type: " + shape.getType()
+      );
+    };
+  }
+
+  @Override
+  protected TokenTree toDafny(
+    final Shape shape,
+    final String rustValue,
+    boolean isRustOption,
+    boolean isDafnyOption
+  ) {
+    return switch (shape.getType()) {
+      case STRING, ENUM -> {
+        if (shape.hasTrait(EnumTrait.class) || shape.isEnumShape()) {
+          var enumShapeName = toSnakeCase(shape.toShapeId().getName());
+          if (isDafnyOption) {
+            yield TokenTree.of(
+              """
+              ::std::rc::Rc::new(match &%s {
+                  Some(x) => crate::_Wrappers_Compile::Option::Some { value: crate::conversions::%s::to_dafny(x.clone()) },
+                  None => crate::_Wrappers_Compile::Option::None { }
+              })
+              """.formatted(rustValue, enumShapeName)
+            );
+          } else if (isRustOption) {
+            yield TokenTree.of(
+              "crate::conversions::%s::to_dafny(%s.clone().unwrap())".formatted(
+                  enumShapeName,
+                  rustValue
+                )
+            );
+          } else {
+            yield TokenTree.of(
+              "crate::conversions::%s::to_dafny(%s.clone())".formatted(
+                  enumShapeName,
+                  rustValue
+                )
+            );
+          }
+        } else if (shape.hasTrait(DafnyUtf8BytesTrait.class)) {
+          final String rustToDafny =
+            "dafny_runtime::dafny_runtime_conversions::vec_to_dafny_sequence(&%s.as_bytes().to_vec(), |b| *b)";
+          String valueToDafny;
+          if (isRustOption) {
+            valueToDafny =
+              """
+              match %s {
+                Some(s) => crate::_Wrappers_Compile::Option::Some { value: %s },
+                None => crate::_Wrappers_Compile::Option::None {},
+              }""".formatted(rustValue, rustToDafny.formatted("s"));
+            if (!isDafnyOption) {
+              valueToDafny = "(%s).Extract()".formatted(valueToDafny);
+            }
+          } else {
+            valueToDafny = rustToDafny.formatted(rustValue);
+          }
+          yield TokenTree.of("::std::rc::Rc::new(%s)".formatted(valueToDafny));
+        } else {
+          if (isRustOption) {
+            var result = TokenTree.of(
+              "crate::standard_library_conversions::ostring_to_dafny(&%s)".formatted(
+                  rustValue
+                )
+            );
+            if (!isDafnyOption) {
+              result = TokenTree.of(result, TokenTree.of(".Extract()"));
+            }
+            yield result;
+          } else {
+            yield TokenTree.of(
+              "dafny_runtime::dafny_runtime_conversions::unicode_chars_false::string_to_dafny_string(&%s)".formatted(
+                  rustValue
+                )
+            );
+          }
+        }
+      }
+      case BOOLEAN -> {
+        if (isRustOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::obool_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        } else {
+          yield TokenTree.of("%s.clone()".formatted(rustValue));
+        }
+      }
+      case INTEGER -> {
+        if (isDafnyOption) {
+          if (isRustOption) {
+            yield TokenTree.of(
+              "crate::standard_library_conversions::oint_to_dafny(%s)".formatted(
+                  rustValue
+                )
+            );
+          } else {
+            yield TokenTree.of(
+              "crate::standard_library_conversions::oint_to_dafny(Some(%s))".formatted(
+                  rustValue
+                )
+            );
+          }
+        } else {
+          yield TokenTree.of("%s.clone()".formatted(rustValue));
+        }
+      }
+      case LONG -> {
+        if (isRustOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::olong_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        } else {
+          yield TokenTree.of("%s.clone()".formatted(rustValue));
+        }
+      }
+      case DOUBLE -> {
+        if (isRustOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::odouble_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        } else {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::double_to_dafny(*%s)".formatted(
+                rustValue
+              )
+          );
+        }
+      }
+      case TIMESTAMP -> {
+        if (isRustOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::otimestamp_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        } else {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::timestamp_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        }
+      }
+      case BLOB -> {
+        if (isDafnyOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::oblob_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        } else if (isRustOption) {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::oblob_to_dafny(&%s).Extract()".formatted(
+                rustValue
+              )
+          );
+        } else {
+          yield TokenTree.of(
+            "crate::standard_library_conversions::blob_to_dafny(&%s)".formatted(
+                rustValue
+              )
+          );
+        }
+      }
+      case LIST -> {
+        ListShape listShape = shape.asListShape().get();
+        Shape memberShape = model.expectShape(
+          listShape.getMember().getTarget()
+        );
+        if (!isDafnyOption) {
+          if (isRustOption) {
+            yield TokenTree.of(
+              """
+              ::dafny_runtime::dafny_runtime_conversions::vec_to_dafny_sequence(&%s.clone().unwrap(),
+                  |e| %s,
+              )
+              """.formatted(rustValue, toDafny(memberShape, "e", false, false))
+            );
+          } else {
+            yield TokenTree.of(
+              """
+              ::dafny_runtime::dafny_runtime_conversions::vec_to_dafny_sequence(&%s,
+                  |e| %s,
+              )
+              """.formatted(rustValue, toDafny(memberShape, "e", false, false))
+            );
+          }
+        } else {
+          yield TokenTree.of(
+            """
+            ::std::rc::Rc::new(match &%s {
+                Some(x) => crate::r#_Wrappers_Compile::Option::Some { value :
+                    ::dafny_runtime::dafny_runtime_conversions::vec_to_dafny_sequence(x,
+                        |e| %s,
+                    )
+                },
+                None => crate::r#_Wrappers_Compile::Option::None {}
+            })
+            """.formatted(rustValue, toDafny(memberShape, "e", false, false))
+          );
+        }
+      }
+      case MAP -> {
+        MapShape mapShape = shape.asMapShape().get();
+        Shape keyShape = model.expectShape(mapShape.getKey().getTarget());
+        Shape valueShape = model.expectShape(mapShape.getValue().getTarget());
+        if (!isDafnyOption) {
+          if (isRustOption) {
+            yield TokenTree.of(
+              """
+              ::dafny_runtime::dafny_runtime_conversions::hashmap_to_dafny_map(&%s.clone().unwrap(),
+                  |k| %s,
+                  |v| %s,
+              )
+              """.formatted(
+                  rustValue,
+                  toDafny(keyShape, "k", false, false),
+                  toDafny(valueShape, "v", false, false)
+                )
+            );
+          } else {
+            yield TokenTree.of(
+              """
+              ::dafny_runtime::dafny_runtime_conversions::hashmap_to_dafny_map(&%s.clone(),
+                  |k| %s,
+                  |v| %s,
+              )
+              """.formatted(
+                  rustValue,
+                  toDafny(keyShape, "k", false, false),
+                  toDafny(valueShape, "v", false, false)
+                )
+            );
+          }
+        } else {
+          yield TokenTree.of(
+            """
+
+            ::std::rc::Rc::new(match &%s {
+                Some(x) => crate::r#_Wrappers_Compile::Option::Some { value :
+                    ::dafny_runtime::dafny_runtime_conversions::hashmap_to_dafny_map(x,
+                        |k| %s,
+                        |v| %s,
+                    )
+                },
+                None => crate::r#_Wrappers_Compile::Option::None {}
+            })
+            """.formatted(
+                rustValue,
+                toDafny(keyShape, "k", false, false),
+                toDafny(valueShape, "v", false, false)
+              )
+          );
+        }
+      }
+      case STRUCTURE, UNION -> {
+        var structureShapeName = toSnakeCase(shape.getId().getName());
+        if (!isDafnyOption) {
+          if (isRustOption) {
+            yield TokenTree.of(
+              """
+              crate::conversions::%s::to_dafny(&%s.clone().unwrap())
+              """.formatted(structureShapeName, rustValue)
+            );
+          } else {
+            yield TokenTree.of(
+              """
+              crate::conversions::%s::to_dafny(%s.clone())
+              """.formatted(structureShapeName, rustValue)
+            );
+          }
+        } else {
+          yield TokenTree.of(
+            """
+            ::std::rc::Rc::new(match &%s {
+                Some(x) => crate::_Wrappers_Compile::Option::Some { value: crate::conversions::%s::to_dafny(x.clone()) },
+                None => crate::_Wrappers_Compile::Option::None { }
+            })
+            """.formatted(rustValue, structureShapeName)
+          );
+        }
+      }
+      default -> throw new IllegalArgumentException(
+        "Unsupported shape type: %s".formatted(shape.getType())
       );
     };
   }
