+++ codegen/smithy-dafny-codegen/src/main/resources/templates/runtimes/rust/client/operation_builder.rs
@@ -1,12 +1,12 @@
 impl super::Client {
-    /// Constructs a fluent builder for the [`$operationName:L`](crate::operation::$snakeCaseOperationName:L::builders::$operationName:LFluentBuilder) operation.
+    /// Constructs a fluent builder for the [`$operationName:L`](crate::operation::$snakeCaseOperationName:L::builders::$pascalCaseOperationName:LFluentBuilder) operation.
     ///
     /// - The fluent builder is configurable:
 $builderSettersDoc:L
-    /// - On success, responds with [`$operationOutputName:L`](crate::operation::$snakeCaseOperationName:L::$operationOutputName:L) with field(s):
+    /// - On success, responds with [`$pascalCaseOperationOutputName:L`](crate::operation::$snakeCaseOperationName:L::$pascalCaseOperationOutputName:L) with field(s):
 $outputDoc:L
-    /// - On failure, responds with [`SdkError<$operationErrorName:L>`](crate::operation::$snakeCaseOperationName:L::$operationErrorName:L)
-    pub fn $snakeCaseOperationName:L(&self) -> crate::operation::$snakeCaseOperationName:L::builders::$operationName:LFluentBuilder {
-        crate::operation::$snakeCaseOperationName:L::builders::$operationName:LFluentBuilder::new(self.clone())
+    /// - On failure, responds with [`SdkError<$pascalCaseOperationErrorName:L>`](crate::operation::$snakeCaseOperationName:L::$pascalCaseOperationErrorName:L)
+    pub fn $snakeCaseOperationName:L(&self) -> crate::operation::$snakeCaseOperationName:L::builders::$pascalCaseOperationName:LFluentBuilder {
+        crate::operation::$snakeCaseOperationName:L::builders::$pascalCaseOperationName:LFluentBuilder::new(self.clone())
     }
 }
